==> 4. Curva C-R/3.Suavizamiento.ipynb <==
import pandas as pd
import os
import random

df = pd.read_excel('salidas/03-ModeloFinal.xlsx')
df

df000 = df[['pm25_pred','HR']].copy()
df00 = df000[df000['pm25_pred']>=15].copy()
df00 = df000[~((df000['pm25_pred']<=23) &  (df000['HR']>= 1.4))]
df00 = df00[~((df00['pm25_pred']>23) &  (df00['HR']<= 1.0))]
df00.sort_values(by = 'pm25_pred',inplace = True,ignore_index =True)
#df00 = df000.copy()
df00

df_final = pd.DataFrame(columns = df00.columns.to_list())
df_final

pm25_min = df00['pm25_pred'][0]
pm25_max = df00['pm25_pred'][len(df00)-1]
n = 5
m = 4
salto = (pm25_max-pm25_min)/n
salto

pm25 = pm25_min
for pm25 in range(int(pm25_min),int(pm25_max)):
    ####SELECCION DE DATOS
    df_grupo = df00[(df00['pm25_pred'] >=pm25) & 
                    (df00['pm25_pred'] <pm25+salto)].copy()
    ###ASIGNAR INDICES DE FORMA ALEATORIA 
    s = []
    for x in range(0,len(df_grupo)):
        s.append(random.randint(0,len(df_grupo)))
    df_grupo['orden'] = s
    df_grupo.sort_values('orden',inplace = True,ignore_index=True)
    df_grupo = df_grupo[['pm25_pred','HR']].copy()
    k = int(len(df_grupo)/m)
    pm25 += salto
    
    for i in range(0,len(df_grupo)):
        if i%k == 0:
        #print(i)
        #df1 = df10[i:i+k].groupby('pm25_pred').median().reset_index()
            df0 = df_grupo[i:i+k]
            df1 = pd.DataFrame({df00.columns.to_list()[0] : [df0.pm25_pred.mean()], 
                   df00.columns.to_list()[1] : [df0.HR.median()]})
            df_final = pd.concat([df_final,df1],ignore_index=True)
        #print(df1)
        else:
            continue

print(len(df_final))
df_final = df_final[df_final['HR']<1.8]
df_final = df_final[df_final['HR']>0.8]
df_final.sort_values(by = 'pm25_pred',inplace = True,ignore_index =True)
minpm25 = df_final['HR'][0]
df_final['HR'] = df_final['HR']/minpm25
print(len(df_final))


#df_final.to_excel('salidas/Pruebas/06-Promedio.xlsx',index=False)
df_final.to_excel('salidas/04-Promedio.xlsx',index=False)
